Ref #12 Any remote change in secondary connection will result in resource update
Combining *Optional* and *Computed* in a Set type is not behaving properly.
If parameters are not defined in a resource, hash is computed with empty values and plan is not clean when the values come from remote.

Solution was applied so before setting value from remote, there is a check if there is existing value, if no then remote is ignored (ignoring Compute behaviour)