Use `metal-go` for the device data source
This updates the device data source to use `metal-go` instead of
`packngo`.  As part of this migration, the device data source
gains the new `sos_hostname` attribute which enables customers
to reduce their reliance on the deprecated `facility` attribute,
which was previously necessary in order to compute the SOS
hostname for a device.