Allow custom_data and user_data to change without ForceNew
By default, we force recreation of Equinix Metal devices when the
`user_data` or `custom_data` attributes change, but this decision
only exists within the terraform provider; the Equinix Metal API
will happily accept updates to those fields without deleting a
device.

Previously we had narrow support to bypass this behavior with a
`reinstall` block, but that still involves downtime.  This adds
support for `behavior.allow_changes`, which is a list of attributes
that are allowed to change _without_ forcing device recreation.

Closes #159