Specify full import path to the ProviderVersion constant
We have not been properly injecting the release version into the
ProviderVersion constant, so we cannot see terraform provider usage
by version at the moment.  This updates the goreleaser config to
specify the full import path to the ProviderVersion constant so
that the release version ends up where we need it.