ci: Stabilize plan/metro/facility selection in tests
This adds the `random` provider to our acceptance tests so that we
can use the `random_integer` resource to pick a random plan, facility
and metro in each test and persist that choice in state.

By persisting the first random selection in state for each test, we
avoid the potential to make a different random choice for each step
in a multi-step test.  This enables us to remove `lifecycle` blocks
from our tests so that we can more easily test the full behavior of
our resources.