fix: don't set empty vnid before deleting virtual circuit
The previous fix (removing unnecessary `port` from the VC state
waiter) uncovered a downstream API error in the virtual circuit
delete code when deleting a VRF virtual circuit.

When any virtual circuit is deleted, the terraform provider was
trying to disconnect the VLAN from the circuit. This request was
getting a 422 response from the API with the message "Update
params can't be blank". Running locally, both the VRF and VLAN
virtual circuit tests pass without this explicit disconnect
request.