ci: use spot market price as max bid in tests
The spot market resource tests use a data source to determine the
spot market price, and then attempt to create a spot market request
with a bid price that is 20% higher than the spot market price.

This introduces brittleness to our tests because the bid price may
be higher than the maximum allowed bid price.  This change uses the
spot market price as the max bid price, rather than setting a higher
bid, in an attempt to reduce the chance of test failure due to a
disallowed bid price.