chore(ci): fix SSH key test and document project name max length
At some point, the Metal API introduced validation that rejects project
names longer than 80 characters.  This broke one of our device tests,
which was using the public SSH key contents as the project suffix instead
of a random number.

This updates the broken test to use a random number suffix for project
names so that the test can pass again.  The project resource docs are also
updated to mention the 80-character limit.