chore(ci): convert to a push-button release workflow
Currently, the release workflow requires someone to manually inspect
the changes since the last release, decide on a good version number
to use, create that tag locally, and push it to GitHub to trigger a
release.

Instead, we could follow the pattern established with our generated
SDKs and use Conventional Commit tags to enable automatic calculation
of the next version number.  This enables us to release with a single
click, which makes it easier to release quickly and share responsibility
for creating releases.