ci: run acctest workflows 1 at a time and support external PRs (#374)
We don't need to hammer the API with end-to-end tests when the repo gets
a flurry of activity. This adds a concurrency setting to our acceptance
test workflow with a hard-coded string value so that only one run of the
acceptance tests can execute at any time.

This also updates the acceptance test workflow to replace the `pull`
trigger with a `pull_request_target` trigger. We've been using this for
`metal-cli` and it has been a good experience thus far. The new trigger
allows us to manually run end-to-end tests for PRs from forks in order
to build confidence in code changes before they are merged.