fix: read metal-go client's auth token from config, not env
When the metal-go client was introduced to enable packngo deprecation,
it was wired up directly to the `METAL_AUTH_TOKEN` environment variable
for auth.  This breaks auth when a customer is using the `auth_token`
provider configuration.

This updates the metal-go client initialization so that the client will
work with both the `auth_token` setting and the `METAL_AUTH_TOKEN` env
var.