fix: use the correct base User-Agent for metal-go requests
The code that injects terraform module information into the User-Agent
string for the metal-go client was mis-wired; it was using the packngo
User-Agent as a base instead of the metal-go User-Agent.  This change
will correct the User-Agent so that requests made with metal-go will
not appear to come from packngo.