refactor: clean up config package
When the config struct was moved to its own package, we added a couple
public variables to replace private ones.  However, the original private
variable definitions were left in place, and all references to the private
variables were left unchanged.

On closer inspection, one of the private variables is only used by one
resource and can remain private, and the other one is for validation, not
configuration.  This moves the validation variable to a validation package
and removes the unused variables.