refactor: clean up config package (#480)
When the config struct was moved to its own package, we added a couple
public variables to replace private ones. However, the original private
variable definitions were left in place, and all references to the
private variables were left unchanged.

On closer inspection, one of the private variables (`ipAddressTypes`) is
only used by one resource and can remain private. The other one
(`uuidRE`) is used for validation that duplicates a built-in terraform
validator, so I've replaced it with the built-in validator.