refactor: remove unnecessary wrappers for getting a random string or int (#481)
Some of the Network Edge tests were using `randInt` and `randString` to
generate random integers and strings. The only thing each function does
is call an upstream function; this replaces all calls to those functions
with direct calls to the upstream (which matches what we do in other
tests) and removes the unnecessary wrappers.