refactor: move spot market timeouts out of provider.go (#483)
There was a private variable for default timeouts defined in
`provider.go`, but it was only referenced in the spot market request
data source. Other resources and data sources define their default
timeouts inline; this does the same for the spot market request data
source so that it will be easier to move that data source to a different
package than the one that `provider.go` is in.