fix: correctly set organization_id in metal_project data source
The metal_project data source was attempting to read the organization
ID for the specified project by looking directly at the `organization.id`
field.  That field is only present on a project when the API request has
`organization` in the `include` parameter.  Rather than include organization
details and make the data source slower, this replicates the logic from
the metal_project resource, which uses the organization's `href` field to
get the organization ID.

While I was in this area of the code I also removed some debugging print
statements that were merged unintentionally.