chore: fix NE path mappings and default owners in CODEOWNERS (#543)
One of the path mappings specified for the Network Edge (NE) team was
`*equinix_network*` which didn't match any files. This replaces that
line with 2 lines to match NE resources and data sources. We could use
`*network*`, but that would catch
`resource_metal_device_network_type.go`.

This also updates the default owners for code in this repo to DevRel
Engineering, since we should be responsible for any code in this repo
that is not specific to a particular service.