refactor: move Fabric acceptance tests to the equinix_test package (#537)
The biggest remaining challenge for moving the SDK provider into a
separate package from the provider resources & data sources is the
acceptance tests. This PR moves the existing Fabric acceptance tests
into the `equinix_test` package so that they are isolated from the
provider code without us needing to move files around. Some private
functions have been made public so that they can continue to be used in
both the customer-facing code and the tests.