chore: update GHA workflows to use hashicorp/setup-terraform action (#551)
Relates to #542 

We've seen an increase in `text file busy` failures during test runs on
some PRs. After searching for that error message in GitHub issues for
other providers, I landed on
https://developer.hashicorp.com/terraform/plugin/sdkv2/testing/acceptance-tests#github-actions-workflow,
which recommends using [the hashicorp/setup-terraform
action](https://github.com/hashicorp/setup-terraform) in GitHub Actions
workflows.

This updates our acceptance test workflows to use the setup action to
install terraform, in the hopes that this reduces or eliminates the
`text file busy` errors.