fix: restore the metal_project datasource name filter (#554)
Closes #549

When the Metal project data source was migrated to `equinix-sdk-go`,
finding a project by name was broken in some cases. To ensure that the
data source can always find the named project (assuming it really
exists), we have to use the `ExecuteWithPagination()` function to load
all pages of projects instead of `Execute()` which only loads one page.

As a minor optimization, this uses the `Name()` filter to reduce the
number of results that are returned by the API.

I've also taken the opportunity to move the Metal project resource and
data source to a separate package.