fix: maintain context in resources and datasources
We recently introduced equinix-sdk-go as a context-aware replacement
SDK for packngo, and _also_ migrated the some resources and datasources
to use context-aware CRUD lifecycle methods.  In those parallel changes
we missed the opportunity to pass the terraform provider context into
the SDK.

This updates the provider to replace all non-testing usage of `context.TODO()`
and  `context.Background()` with references to the pre-existing terraform
provider context.