chore: capture test coverage for unit tests (#569)
I'm pretty sure we were capturing this previously because all tests
(unit and integration) were running in the same workflow. In any case,
this captures the unit test coverage because it's at least as useful to
track as the integration test coverage.

From what I can see in CodeCov docs, the unit and integration reports we
upload for a particular commit will be merged automatically into one
report, so we'll get a better picture of how well our code is covered.

This also updates the unit test workflow to trigger on every push to
main; since these tests are lightweight and siolated, it's safe to run
them frequently. And it ensures that our test status badge in the README
stays up-to-date!