chore(ci): skip shared connections when sweeping virtual circuits (#584)
The sweeper step of our end-to-end test workflow sometimes fails because
it finds & attempts to delete virtual circuits that belong to shared
interconnections. The API blocks deletion of these circuits because they
are API-managed, not user-managed, and can only be deleted by deleting
the interconnection itself. This adds a check to the virtual circuit
sweeper so that the sweeper will ignore virtual circuits that belong to
shared interconnections.