chore(ci): update test provider wiring for resources under internal/
A previous PR updated "all" Metal resource & datasource acceptance tests
to use ProtoV5ProviderFactories instead of Provider so that we don't need
to update unrelated tests whenever we migrate a resource from the SDK
provider to the framework provider.

That PR missed a few tests; this updates those tests and removes some
test provider structures that are no longer needed.