chore(ci): update sweeper registration for Metal resources (#593)
When we migrated some resources to separate packages under
`internal/resources`, the sweepers for the resources that were moved
stopped running. This adds a new `internal/sweeper` package that sets up
the sweepers for resources under `internal/resources`.

In most cases, that change would be enough by itself, but the Metal
sweepers have some interdependencies, and those interdependencies only
work correctly if every relevant sweeper is registered within the same
TestMain function. To resolve that, thi moves the test sweepers
for Metal resources that are still in `equinix/` from the test file to
the main resource file and adds an `equinix.AddTestSweepers` shim that
allows the sweeper in `internal/sweep` to know about all Metal sweepers
regardless of location.