feat: push connection speed validation down to the API (#610)
Previously, we maintained a list of valid connection speed values in the
terraform provider in order to steer users towards selecting a valid
connection speed. As shown by #595, validating values in the provider
has the potential to break when the list of connection speed values
supported by the API changes.

This updates the code that converts speed values between integer and
string representations (i.e., the code that knows that `"50Mbps"` is the
same as `500000000`) to convert (roughly) arbitrary values. This means
that new values are immediately "supported" by the provider, but it also
means that invalid values will not be flagged until apply.

The test change in this PR is sufficient to resolve the issue introduced
by merging #605 without any other change to provider behavior.

Closes #609