fix: convert metal_connection delete to equinix-sdk-go
The metal_connection resource and data source were previously migrated
from the deprecated packngo SDK to equinix-sdk-go, but in that migration
the Delete function for the metal_connection resource was missed.

This updates the resource Delete function to use equinix-sdk-go instead
of packngo.  This requires adding a state waiter to replicate the wait
functionality that was built in to packngo.  As part of this migration,
the waiter is updated to allow the `deleted` status that can be returned
by the API for a deleted connection.