chore: allow continued use of stop context for now (#737)
The linter will complain if any changes are made in
`equinix/provider.go` because we are using a deprecated stop context for
coordinated shutdown across resources. We have to continue using a stop
context until all resources and data sources are updated to use
context-aware CRUD functions. This disables staticcheck for the relevant
line of code and adds a comment explaining why linting was disabled in
this case.