feat!: switch from tfprotov5 to tfprotov6 (#806)
Closes #738 

This adopts [Terraform Plugin Protocol version
6](https://developer.hashicorp.com/terraform/plugin/terraform-plugin-protocol#protocol-version-6)
so that this provider can leverage newer features of the protocol,
enabling provider developers to easily reference the latest plugin
development docs and provide a better end-user experience.

Upgrading the Terraform Plugin Protocol means dropping support for
Terraform < 1.0, which constitutes a breaking change. No resources or
datasources have been updated in this PR to take advantage of the
features enabled by the latest protocol. The new protocol features are
only available to framework-based resources and datasources, and as of
this PR, only Metal has those.