chore(ci): delay release publish until after artifacts are attached (#834)
We have intermittently seen provider releases fail to show up in the
Terraform registry. When that happens we can usually redeliver the
GitHub release webhook event and get the registry to find the new
release.

While working on changes to the Equinix Python SDK, it occurred to me
that the webhook issues we are seeing might be related to how we do
automatic releases. The semantic-release-action creates the GitHub
release, and then the provider artifacts are built and uploaded by
GoReleaser after the GitHub release has been created. The webhook event
is sent to the Terraform registry when the release is created, not after
the artifacts are uploaded; it's possible that the registry attempts to
load the release only to find that it does not include a provider, and
that that is why our provider releases sometimes do not show up in the
registry right away.

This PR updates the semantic-release configuration so that
semantic-release-action creates a draft release, and the release is not
published until after GoReleaser builds and uploads the necessary
artifacts. The release and publish steps are split into separate jobs in
the release workflow so that we can more easily re-run specific parts of
the workflow and easily skip the publish step if no new release was
created. These changes may resolve our issues with releases not showing
up in the registry, but even if they don't, this is still a worthwhile
quality-of-life improvement for the release workflow.