chore: update local lint task to only lint changed files (#851)
Since we might periodically change linter configuration, as done in
#849, we should keep the existing PR validation behavior that only lints
the files that were changed in a particular PR. This will allow us to
continue to avoid unnecessarily large PRs to fix linting failures if and
when the linter configuration changes.

If we're keeping the current, limited PR validation behavior, we should
use the same behavior for the `make lint` task to that developers can
easily identify the minimum necessary changes locally. This updates the
`make lint` task to only lint files that have been changed since
`origin/main`, which should closely approximate the behavior of the PR
validation workflow.