chore: allow underscores in names in test files (#856)
HashiCorp examples encourage the use of underscores in names in
Terraform provider tests. To avoid unnecessary churn due to renaming
things in test files, we disabled the relevant `revive` rule for tests,
but it turns out there's a similar `stylecheck` rule that also needs to
be skipped. This PR disables that rule for test files.