chore: use original API request context for auth token requests (#847)
#844 updated the `Config` object to create a new auth client every time
a new Fabric API client was created for SDK resources. If we stuck with
this pattern, we would have to duplicate that behavior for every Equinix
API client except for Metal. In addition, this pattern means that the
provider will make a separate auth token request per resource rather
than reusing valid tokens across all resources that use the same
provider (and therefore auth) config.

This reverts the auth client change by adopting a custom HTTP transport
and OAuth token source that enables passing the original request context
for an API request to the auth token client. As a result, we no longer
need to use `equinix/oauth2-go`.