refactor: Update equinix_metal_port_vlan_attachment to use equinix-sdk-go (#975)
Equinix-sdk-go is the preferred library for interacting with the Metal
API now. In this PR I've updated the
`equinix_metal_port_vlan_attachment` resource so that it is using that
SDK rather than the older packngo SDK.

The equinix-sdk-go interfaces expect a context, so I also updated the
resource to provide functions for the `{C,R,U,D}Context:` values of
`schema.Resource`. The previous functions did not accept a context.

Eventually I'll want to move away from `Assign/Unassign` endpoints here,
but this change helps make the future change smaller.