fix: Handle equinix_metal_port updates properly if no changes to VLANs
In #976, there was a bug introduced when refactoring the batch logic
to be shared between the metal_port and metal_port_vlan_attachment
resources.

The previous code for the port resource always called the batch
endpoints but was a no-op if there were no assignments. I've updated
the VLAN Batch logic to do the same here.